include tick count in all tlm packets
The tick count is now included in all telemetry packets so that all data
can be correlated with a given tick.

Increased telemetry packet buffer size from 16 to 32 so that the tick
count can fit.

+++ src/hardware_model/ballast/mod.rs
@@ -99,7 +99,7 @@ impl Ballast {
 }
 
 impl Tick for Ballast {
-    fn tick(&mut self, _tick_count: u128) {
+    fn tick(&mut self, _tick_count: u32) {
         if self.state == BallastState::Transition {
             self.stop_all();
         }
